Create class `LinearWebCollector`
